* fix/ILUB-111: (46 commits)
  Working Version
  WIP
  removed static public function
  improved example
  added example
  implemented feature
  FIX BUG User in status ignored (no update at the last time) should also deleted if they are not in the data set the next time
  part 1
  Removed IMetadataFactory/getDTOWithFirstIliasIdForTitle
  code format for demoUser and removed deletion
  code reformat in demoCategory and removed deletion of objects
  fixed old implementation of ignoring deletion
  temporary-fix for taxanomies
  fixed bug when deleting origin
  renamed parameter II
  renamed parameter
  composer fixed
  improvmeent of NULL-DTOs for Hooks
  possiblility to override status on origin
  temporary-fix for taxanomies
  ...

+++ src/Sync/Processor/CourseMembership/CourseMembershipSyncProcessor.php
@@ -18,6 +18,7 @@ use SRAG\Plugins\Hub2\Origin\Properties\CourseOriginProperties;
 use SRAG\Plugins\Hub2\Sync\IObjectStatusTransition;
 use SRAG\Plugins\Hub2\Sync\Processor\FakeIliasMembershipObject;
 use SRAG\Plugins\Hub2\Sync\Processor\ObjectSyncProcessor;
+use SRAG\Plugins\Hub2\Origin\Properties\CourseMembershipOriginProperties;
 
 /**
  * Class CourseMembershipSyncProcessor
@@ -80,7 +81,7 @@ class CourseMembershipSyncProcessor extends ObjectSyncProcessor implements ICour
 		 * @var CourseMembershipDTO $dto
 		 */
 		$obj = FakeIliasMembershipObject::loadInstanceWithConcatenatedId($ilias_id);
-		$ilias_course_ref_id = $dto->getCourseId();
+		$ilias_course_ref_id = $obj->getContainerIdIlias();
 		$user_id = $dto->getUserId();
 		if (!$this->props->updateDTOProperty('role')) {
 			return new FakeIliasMembershipObject($ilias_course_ref_id, $user_id);
@@ -106,6 +107,11 @@ class CourseMembershipSyncProcessor extends ObjectSyncProcessor implements ICour
 	 */
 	protected function handleDelete($ilias_id) {
 		$obj = FakeIliasMembershipObject::loadInstanceWithConcatenatedId($ilias_id);
+
+		if ($this->props->get(CourseMembershipOriginProperties::DELETE_MODE)
+				== CourseMembershipOriginProperties::DELETE_MODE_NONE) {
+			return $obj;
+		}
 
 		$course = $this->findILIASCourse($obj->getContainerIdIlias());
 		$course->getMembersObject()->delete($obj->getUserIdIlias());
